+++ src/main/java/br/com/meli/dhprojetointegrador/repository/AgentRepository.java
@@ -3,5 +3,5 @@ package br.com.meli.dhprojetointegrador.repository;
 import br.com.meli.dhprojetointegrador.entity.Agent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentRepository extends JpaRepository<Agent, Integer>  {
+public interface AgentRepository extends JpaRepository<Agent, Long>  {
 }
