+++ src/main/java/br/com/meli/dhprojetointegrador/repository/SectionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SectionRepository extends JpaRepository<Section, Integer> {
+public interface SectionRepository extends JpaRepository<Section, Long> {
 }
