+++ src/main/java/br/com/meli/dhprojetointegrador/repository/SellerRepository.java
@@ -3,5 +3,5 @@ package br.com.meli.dhprojetointegrador.repository;
 import br.com.meli.dhprojetointegrador.entity.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerRepository extends JpaRepository<Seller, Integer> {
+public interface SellerRepository extends JpaRepository<Seller, Long> {
 }
